+++ Manero/Helpers/Services/ProductService.cs
@@ -66,17 +66,6 @@ namespace Manero.Helpers.Services
         //    return categories.Select(c => c.Name);
         //}
 
-        public async Task<IEnumerable<ProductDetailsEntity>> GetAsync()
-        {
-            var products = await _context.ProductDetails.ToListAsync();
-            return products;
-        }
-
-        public async Task<ProductDetailsEntity> GetAsync(Expression<Func<ProductDetailsEntity, bool>> expression)
-        {
-            var product = await _context.ProductDetails.FirstOrDefaultAsync(expression);
-            return product!;
-
-        }
+     
     }
 }
