+++ Manero/Views/Review/AllReviews.cshtml
@@ -1,5 +1,5 @@
 ﻿@using Manero.Models.Entities;
-@inject ProductService productService
+@inject ProductDetailsService productService
 @inject ReviewService reviewService
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
 @model Manero.Models.Entities.ProductDetailsEntity
